+++ tests/utils/test_utils.go
@@ -5,6 +5,7 @@ import (
 	"encoding/xml"
 	"errors"
 	"fmt"
+	"github.com/jfrog/jfrog-cli-security/formats"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -28,36 +29,15 @@ import (
 	clientTests "github.com/jfrog/jfrog-client-go/utils/tests"
 )
 
-type Vulnerability struct {
-	BomRef string `json:"bom-ref"`
-	Id     string `json:"id"`
-}
-
-type EnrichJson struct {
-	Vulnerability []struct {
-		BomRef string `json:"bom-ref,"`
-		Id     string `json:"id"`
-	} `json:"vulnerabilities"`
-}
-
-type Bom struct {
-	Vulnerabilities struct {
-		Vulnerability []struct {
-			BomRef string `xml:"bom-ref,attr"`
-			Id     string `xml:"id"`
-		} `xml:"vulnerability"`
-	} `xml:"vulnerabilities"`
-}
-
-func UnmarshalJson(t *testing.T, output string) EnrichJson {
-	var jsonMap EnrichJson
+func UnmarshalJson(t *testing.T, output string) formats.EnrichJson {
+	var jsonMap formats.EnrichJson
 	err := json.Unmarshal([]byte(output), &jsonMap)
 	assert.NoError(t, err)
 	return jsonMap
 }
 
-func UnmarshalXML(t *testing.T, output string) Bom {
-	var xmlMap Bom
+func UnmarshalXML(t *testing.T, output string) formats.Bom {
+	var xmlMap formats.Bom
 	err := xml.Unmarshal([]byte(output), &xmlMap)
 	assert.NoError(t, err)
 	return xmlMap
